+++ src/StandardSector.java
@@ -1,5 +1,5 @@
 public class StandardSector extends BaseSector{
-    public StandardSector(int x, int y, MyFrame panel) {
+    public StandardSector(double x, double y, MyFrame panel) {
         super(x, y, panel);
     }
 }
